Create an initial book structure
The aim is to support SPECTRUM and SPARK EMCRs to develop their
computing skills, and to make effective use of available tools and
infrastructure. It may also serve as the basis for a session at the
upcoming SPECTRUM and SPARK EMCR Retreat.

This book is built with mdBook (https://rust-lang.github.io/mdBook/).