Use a Map to define the code block tab names
The CSS class name to tab name mapping is now defined in a Map variable,
which uses the insertion order for iteration, and we use this ordering
to identify the tab name for each code block.

All of the variables and functions in `tabbed-code-blocks.js` are now
defined inside an anonymous function, which acts as a private namespace.