Update mdbook and related dependencies
The updated mdbook adds CSS styling for <kbd> tags, which I want to use
when explaining how to use editors such as nano and vim. I've taken this
opportunity to also update mdbook-admonish, which allows us to fail the
CI builds if an admonition is incorrectly configured, and to update
mdbook-linkcheck (although I suspect this makes no tangible difference).