Use earlier frames for video previews
We previously used the frame at 1 second into the video, but this may
contain a fragment of the first command, so we now use the frame at 0.1
seconds into the video.

We don't use the first frame (0 seconds) because this doesn't include
the initial shell prompt, and instead shows an entirely blank screen.