Divide into several top-level sections
The proposed orientation guide should probably be in a separate section
from the topic-specific materials we've developed so far. To that end,
I've divided the content into the following sections:

- Introduction: as per the original home page, and also containing the
  "How to contribute" and "Contributors" pages.

- Orientation: a new section for the orientation guide.

- Topical guides: the current training materials, except case studies.

- Community: the Community of Practice meetings, and the case studies,
  since they're standalone contributions from the community.

I've also enabled the "navigation.tabs" and "navigation.tabs.sticky"
features, so that these top-level sections are rendered in a menu layer
below the header (except for mobile devices).