Update using-a-debugger.md
Python example of using try - except in debugging. I can write tryCatch() version in R if you think that would be useful. Also, I wanted to add a scheenshot of "the debugging point on line 38" as I don't think that line numbers are visible in the interface but I couldn't add a photo. I am happy to send you the screenshot if you think that would be useful.